Social-seq-live_GUI_Action/requirement.txt
PySide6==6.7.2
pyserial==3.5
pandas
ffmpegcv
